Enhance Jet cleanup logic
In the cleanup logic, if a job completes between cleanup-resources and
cleanup-job-results phase, it is possible that the result is removed
before the resources which leads to a 2 hours delay to destroying the
resources map.

With the changes, we are destroying the resources map when removing the
job-result.

We've introduced a `existsDistributedObject` method to ProxyService to
check if a distributed object exists without creating a lot of litter.